[#] refactor(gateway) : 응답반환 시간에 따른 게이트웨이 오류 처리에 대한 문제 해결
