+++ service/eureka/eureka-core/src/main/java/radiata/service/eureka/core/EurekaApplication.java
@@ -6,9 +6,9 @@ import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
 
 @SpringBootApplication
 @EnableEurekaServer
-public class eurekaApplication {
+public class EurekaApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(eurekaApplication.class, args);
+        SpringApplication.run(EurekaApplication.class, args);
     }
 }
