+++ service/user/user-api/src/main/resources/application.yml
@@ -3,7 +3,7 @@ spring:
     name: user-service
   config:
     import: classpath:application-database.yml
-    
+
 server:
   port: 8083
 
@@ -11,4 +11,6 @@ eureka:
   client:
     service-url:
       defaultZone: http://localhost:19090/eureka/
+  instance:
+    hostname: localhost
